Add basic edited messages handling
When the user edits a vote message so that its date is now set to the future
and voting is possible, the bot can handle it even if it was in an invalid state before,
e.g. it was set to a past datetime. It won't restart the timer and add the reactions
it should have, but if a user adds the reaction, it adds it as well, so it stays there