Rework rolehoarders_find
Make the command name shorter by changing it to "rolehoarders".

Change how the optional parameter and config value works to make it more
logical. The given number now represents the number of roles a user has
to have to be considered a rolehoarder (and therefore listed in the
command output).

Change the output format of the command, the users are now tagged. Also
allow the command to be run outside of modroom but only allow Helper+ to
use the command.

Signed-off-by: František Nečas <frantisek.necas@protonmail.com>